requirements/base.txt
django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
markdown==3.8
django-filter==25.1
django-cors-headers==4.7.0
drf-spectacular==0.28.0
python-decouple==3.8
psycopg2-binary==2.9.10
Faker==37.1.0